% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyEncoding.R
\name{applyEncoding}
\alias{applyEncoding}
\title{applyEncoding}
\usage{
applyEncoding(obj, ...)
}
\arguments{
\item{obj}{an encoding object created by one of the encoding functions in this package}

\item{...}{other arguments dependent on object class}
}
\description{
applyEncoding
}
