% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencyEncoding.R
\name{frequencyEncode}
\alias{frequencyEncode}
\title{frequencyEncode}
\usage{
frequencyEncode(dt, vars, encodeNA = FALSE, allowNewLevels = FALSE)
}
\arguments{
\item{dt}{data.frame(table) to create the object on}

\item{vars}{vector of variables you want to frequency-encode}

\item{encodeNA}{Boolean. Should NAs be encodes as a frequency, or kept as NA when the transformation is applied? If
there are no NAs in your original data, new NAs will still be encoded as 1. Risky, but easy.}

\item{allowNewLevels}{Should any new levels be encoded as -1? Details:
\itemize{
  \item \code{TRUE}   Encodes new levels as -1. This is dangerous if your levels can change in the future, because you won't notice, and the model may not be tuned correctly.
  \item \code{FALSE}  Throws an error. You'll need to figure out how you want to proceed if \code{allowNewLevels = TRUE} is not good enough.
}}
}
\value{
Frequency Encoded Object. This needs to be applied to a dataset, it will not actually return a dataset.
}
\description{
Re-encodes categorical data as its frequency in the dataset. Useful for gradient boosting.
Does NOT return dataset, but an object that can be applied to a dataset with the \code{applyEncoding} function.
If your data contains missing values, be very careful with the \code{encodeNA} and \code{allowNewLevels} parameters.
}
