% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianEncoding.R
\name{gaussianEncode}
\alias{gaussianEncode}
\title{gaussianEncode}
\usage{
gaussianEncode(dt, vars, newMean = 0, newSD = 1, encodeNA = NA)
}
\arguments{
\item{dt}{Dataset to create object on.}

\item{vars}{variables you want to include in the encoding.}

\item{newMean}{The mean of the transformed distribution}

\item{newSD}{The SD of the transformed distribution}

\item{encodeNA}{You can impute NAs as a manual value, or leave them as NA.}
}
\value{
Rare Encoded Object
}
\description{
gaussianEncode
}
