% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odbc.R
\name{getQuery}
\alias{getQuery}
\title{getQuery}
\usage{
getQuery(
  email,
  driver = "ODBC Driver 17 for SQL Server",
  server,
  database,
  sql
)
}
\arguments{
\item{email}{Whatever UID you sign into SQL server with.}

\item{driver}{Try leaving empty first.
You can use ```odbc::odbcListDrivers()``` to see which
drivers you have ifthe default doesn't work.}

\item{server}{The server}

\item{database}{Need to specify a connection to the database.}

\item{sql}{SQL string to run.}
}
\value{
data.table results from sql
}
\description{
Placeholder
}
