% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naEncoding.R
\name{naEncode}
\alias{naEncode}
\title{naEncode}
\usage{
naEncode(
  dt,
  vars = names(dt),
  naReplaceNumber = NULL,
  naReplaceCharacter = NULL,
  inPlace = TRUE,
  verbose = T
)
}
\arguments{
\item{dt}{dataset of class `data.table` or `data.frame`}

\item{vars}{The variables in dt you want to impute. If NULL, uses all columns.}

\item{naReplaceNumber}{replacement number for numerics.}

\item{naReplaceCharacter}{replacement string for characters and factors.}

\item{inPlace}{should the entire dataset be returned, or just the variables in \code{vars}}

\item{verbose}{Print results?}
}
\description{
Replaces NA and NaN values in \code{dt[vars]}. Will add a new level to factors.
}
