% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityPlotting.R
\name{plot1dDensity}
\alias{plot1dDensity}
\title{plot2dContour}
\usage{
plot1dDensity(dat, x, grouping)
}
\arguments{
\item{dat}{the data.table(frame) that contains the data.}

\item{x}{The variable on the x axis. Should be a character.}

\item{grouping}{Will create multiple contours broken up by this variable. Should be a character name of the column you want to use.}
}
\value{
a ggplot.
}
\description{
Plots the density of a variable, with optional grouping.
}
