% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_samples.R
\name{prepare_samples}
\alias{prepare_samples}
\title{Prepare samples}
\usage{
prepare_samples(path, patt, trimmed.reads)
}
\arguments{
\item{path}{List of full paths to directory or directories containing the raw reads data}

\item{patt}{List of suffix patterns for the raw reads data for forward and (optionally) reverse reads}

\item{trimmed.reads}{Name of the directory for the quality and adapter trimmed reads, optional}
}
\value{
Dataframe with sorted file paths and sample names
}
\description{
Reads in the sample file names for the directory or directories. Ensures the files are in order, returns a dataframe with the paths
             to the raw reads files, the trimmed reads files and the sample names. Can take paired end or single end data.
             If there are duplicated sample names the files will be merged into a temporary directory.
             For scRNA samples, require paired end reads and do not require the trimmed reads directory. Call the function prepare_solo_samples.
}
\examples{
\dontrun{
# Paths to reads
reads.path <- c("/Path to reads 1",
                "/Path to reads 2")

# Standard suffixes for reads files for forward and reverse files
reads.patt.1 <- "*_R1_001.fastq.gz$"
reads.patt.2 <- "*_R2_001.fastq.gz$"

# Trimmed reads directory
trimmed.reads <- "trimmed_reads"

# Paired end data
sample.dataframe <- prepare_samples(reads.path, c(reads.patt.1,reads.patt.2),trimmed.reads)
mate1 <- as.character(sample.dataframe$reads.path.1)
mate1.trim <- as.character(sample.dataframe$trimmed.reads.path.1)
mate2 <- as.character(sample.dataframe$reads.path.2)
mate2.trim <- as.character(sample.dataframe$trimmed.reads.path.2)

# Single end data
sample.dataframe <- prepare_samples(reads.path, reads.patt.1,trimmed.reads)
mate1 <- as.character(sample.dataframe$reads.path.1)
mate1.trim <- as.character(sample.dataframe$trimmed.reads.path.1)

# Sample names
sample.names <- as.character(sample.dataframe$sample.names)
}

}
