% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_solo_samples.R
\name{prepare_solo_samples}
\alias{prepare_solo_samples}
\title{Title}
\usage{
prepare_solo_samples(path = path, patt)
}
\arguments{
\item{path}{List of full paths to directory or directories containing the raw reads data}

\item{patt}{List of suffix patterns for the raw reads data for forward and reverse reads}
}
\value{
Dataframe with sorted file paths and sample names
}
\description{
Reads in the sample file names for the directory or directories, for scRNA samples. Requires paired end reads.
}
\examples{
reads.path <- c("/Path to reads 1",
                "/Path to reads 2")

# Standard suffixes for reads files for forward and reverse files
reads.patt.1 <- "*_R1_001.fastq.gz$"
reads.patt.2 <- "*_R2_001.fastq.gz$"

sample.dataframe <- prepare_solo_samples(reads.path, c(reads.patt.1,reads.patt.2))

mate1 <- as.character(sample.dataframe$reads.path.1)
mate2 <- as.character(sample.dataframe$reads.path.2)

sample.names <- as.character(sample.dataframe$sample.names)

}
