% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangerFeatureImportance.R
\name{rangerFeatureImportance}
\alias{rangerFeatureImportance}
\title{makeFeatEngTools}
\usage{
rangerFeatureImportance(dat, holdOut, response, runs = 10, plot = TRUE, ...)
}
\arguments{
\item{dat}{The data}

\item{holdOut}{A second dataset to test performance. Not yet implemented.}

\item{response}{The response column}

\item{runs}{How many times to repeat}

\item{plot}{Plot the results}

\item{...}{Other arguments to pass to ranger.}
}
\value{
A matrix of prediction.errors for each run and variable.
}
\description{
Creates and stores functions for feature engineering.
}
