% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareEncoding.R
\name{rareEncode}
\alias{rareEncode}
\title{rareEncode}
\usage{
rareEncode(
  dt,
  vars,
  minPerc = NULL,
  minCount = NULL,
  newString = "rareGroup",
  encodeNA = FALSE,
  allowNewLevels = FALSE
)
}
\arguments{
\item{dt}{Dataset to create object on.}

\item{vars}{variables you want to include in the encoding.}

\item{minPerc}{Minimum percentage of the total population required for a group to be considered 'rare'.
For example, if \code{minPerc = 0.1}, and \code{Gender = 'Female'} only makes up 5% of the dataset, then
\code{Female} would be replaced with the string in \code{newString}.}

\item{minCount}{Can be provided instead of minPerc if you have a better idea of credibility in terms of total
samples required. Levels with counts below this value will be grouped into \code{newstring}.}

\item{newString}{Value you want to group rare levels together as.}

\item{encodeNA}{If FALSE, NAs will remain NA even if they would have been grouped.
If TRUE, NAs will be grouped, but only if they are 'rare'. If they are not rare, NAs will be untouched.}

\item{allowNewLevels}{Similar to frequencyEncode parameter. This groups all new levels in with \code{newString}.}
}
\value{
Rare Encoded Object
}
\description{
Groups rare levels together, where 'rare' is defined by either minPerc or minCount.
}
