% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bcftools.R
\name{run_bcftools}
\alias{run_bcftools}
\title{Run bcftools}
\usage{
run_bcftools(
  command = NULL,
  input = NULL,
  output = NULL,
  regions = NULL,
  parallel = FALSE,
  cores = 4,
  execute = TRUE,
  bcftools = NULL,
  version = NULL
)
}
\arguments{
\item{command}{bcftools command to run, at present can only choose isec, view or sort}

\item{input}{vcf files}

\item{output}{ouput directory or file}

\item{regions}{Bed formatted regions of interest file}

\item{parallel}{Run in parallel, default set to FALSE}

\item{cores}{Number of cores/threads to use for parallel processing, default set to 4}

\item{execute}{Whether to execute the commands or not, default set to TRUE}

\item{bcftools}{Path to the bcftools program, required}

\item{version}{Returns the version number}
}
\value{
A list with the bcftools commands
}
\description{
Run bcftools
}
\examples{
\dontrun{
bcftools.path <- "/software/bcftools-v1.9/bin/bcftools"

bcftools.version <- run_bcftools(bcftool = bcftools.path,
                                 version = TRUE)

bcftools.version[1]

# bcftools intersect
command <- "isec"

vcfs <- c("test1.vcf","test2.vcf")
vcfs <- paste(vcfs, collapse = " ")

out.directory <- "test_directory"



bcftools.cmd <- run_bcftools(command = command,
                             input = vcfs,
                             output = out.directory,
                             execute = FALSE,
                             bcftools = bcftools.path)
bcftools.cmd
}

}
