% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_blast.R
\name{run_blast}
\alias{run_blast}
\title{Run BLAST programs}
\usage{
run_blast(
  input = NULL,
  output = NULL,
  title = NULL,
  database = NULL,
  format = NULL,
  parallel = FALSE,
  cores = 4,
  execute = TRUE,
  blast = NULL,
  version = FALSE
)
}
\arguments{
\item{input}{Fasta formatted file for input to makeblastdb or query input fasta for blast}

\item{output}{Name of output file for blast}

\item{title}{Title for the blast database}

\item{database}{Database type e.g. "nucl" 0r "prot", for makeblastdb or database name for blast}

\item{format}{Output format for blast, numeric e.g 6 is a tabular format}

\item{parallel}{Run in parallel, default set to FALSE}

\item{cores}{Number of cores/threads to use for parallel processing, default set to 4}

\item{execute}{Whether to execute the commands or not, default set to TRUE}

\item{blast}{Path to the BLAST program}

\item{version}{Returns the version number}
}
\value{
A list with the BLAST commands
}
\description{
Runs the BLAST progam and creates BLAST databases
}
\examples{
\dontrun{
path_db <- "/software/blast-v2.13.0/bin/makeblastdb"
path_blast <- "/software/blast-v2.13.0/bin/blastn" or "/software/blast-v2.13.0/bin/blastp"

version <- run_blast(blast = path_blast,
                     version =TRUE)
version[1]

run_blast(input = "sample1.fasta",
          title = "sample_name",
          database = "nucl",
          blast = path_db)

run_blast(input = "cbp1.fasta",
          output = "sample1.blast",
          database = "blastDB",
          format = 6,
          blast = path_blast)
}

}
