% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bwa.R
\name{run_bwa}
\alias{run_bwa}
\title{Run the BWA aligner}
\usage{
run_bwa(command = NULL, mate1 = NULL, mate2 = NULL, index = NULL,
  sample.name = NULL, out.dir = NULL, threads = 10,
  seed.length = NULL, bwa = NULL, version = FALSE)
}
\arguments{
\item{command}{BWA command to run, at present can choose 'mem',
which is most useful fo Illumina, PacBio and Nanopore reads, required}

\item{mate1}{List of the paths to files containing to the forward reads, required}

\item{mate2}{List of the paths to files containing to the reverse reads, required for paired end sequence data}

\item{index}{Path to the indexed reference genome, required}

\item{sample.name}{List of sample names, required}

\item{out.dir}{Name of the directory from the BWA output}

\item{threads}{Number of threads for BWA to use, default set to 10}

\item{seed.length}{Minimum seed length}

\item{bwa}{Path to the bwa program, required}

\item{version}{Returns the version number}
}
\value{
A list with the BWA commands
}
\description{
Runs the BWA aligner
}
\examples{
 \dontrun{
 path <- "/software/bwa/bwa"

# Get version number
bwa.version <- run_bwa(bwa = path,
                       version = TRUE)
bwa.version[3] # print the line with version number

reads.path <- "/path/to/fastqs"
reads.patt.1 <- "*_R1_001.fastq.gz$"
reads.patt.2 <- "*_R2_001.fastq.gz$"

# Organise the samples
sample.dataframe <- prepare_samples(reads.path, c(reads.patt.1,reads.patt.2),trimmed.reads.dir)

mate1 <- as.character(sample.dataframe$reads.path.1)
mate1.trim <- as.character(sample.dataframe$trimmed.reads.path.1)
# Paired end only
 mate2 <- as.character(sample.dataframe$reads.path.2)
 mate2.trim <- as.character(sample.dataframe$trimmed.reads.path.2)

sample.names <- as.character(sample.dataframe$sample.names)

genome <- "/path/to/reference/genome"

command <- "mem"
out.dir <- "bwa_alignments"

# Paired end
pe <- run_bwa(command = command,
              mate1 = mate1.trim,
              mate2 = mate2.trim,
              index = genome,
              threads = 20,
              out.dir = out.dir,
              sample.name = sample.names,
              seed.length = 21,
              bwa = path)

# Single end
se <- run_bwa(command = command,
              mate1 = mate1.trim,
              index = genome,
              threads = 20,
              out.dir = out.dir,
              sample.name = sample.names,
              seed.length = 21,
              bwa = path)
#'  }

}
