% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cellranger.R
\name{run_cellranger}
\alias{run_cellranger}
\title{Run Cellranger}
\usage{
run_cellranger(
  command = NULL,
  input = NULL,
  run.id = NULL,
  description = NULL,
  reference = NULL,
  out.dir = NULL,
  sample = NULL,
  create.bam = "false",
  local.cores = 20,
  local.memory = 96,
  parallel = FALSE,
  cores = 4,
  execute = FALSE,
  cellranger = NULL,
  version = FALSE
)
}
\arguments{
\item{command}{Cellranger command}

\item{input}{Path to input FASTQ directory}

\item{run.id}{A unique run id and output folder name}

\item{description}{Sample description to embed in output files}

\item{reference}{Path of folder containing 10x-compatible reference}

\item{out.dir}{Output the results to this directory}

\item{sample}{Prefix of the filenames of FASTQs to select}

\item{create.bam}{Enable or disable BAM file generation,"true or false", default set to "false"}

\item{local.cores}{Set max cores the pipeline may request at one time, default set to 20}

\item{local.memory}{Set max GB the pipeline may request at one time, defaults set to 96Gb}

\item{parallel}{Run in parallel, default set to "false", quoted in lowercase}

\item{cores}{Number of cores/threads to use for parallel processing, default set to 4}

\item{execute}{Whether to execute the commands or not, default set to TRUE}

\item{cellranger}{Path to the cellranger program, required}

\item{version}{Returns the version number}
}
\value{
A list with the cellranger commands
}
\description{
Runs the Cellranger program, currently only runs cellranger count
}
\examples{
\dontrun{
command <- "count"
fastq_paths <- c("/export/sequencers/nextseq02/processed_data/2026_04/AurelieNajm/3421_VH00574_364_AurelieNajm_20260413/fastq/",
                 "/export/sequencers/nextseq02/processed_data/2026_04/AurelieNajm/3421_VH00574_366_AurelieNajm_20260427/fastq/")
ids <- c("GSYNC-067L","GSYNC-067L","GSYNC-067R","GSYNC-067R","GSYNC-100L","GSYNC-100L","GSYNC-65L","GSYNC-65L","GSYNC-65R",
         "GSYNC-65R","GSYNC-66R","GSYNC-66R","GSYNC-66L","GSYNC-66L","GSYNC-66R","GSYNC-66R","GSYNC-88L","GSYNC-88L","GSYNC-88R","GSYNC-88R")

genome <- "/datastore/Genomes/Homo_sapiens/Ensembl/release-112/indexes/cellranger"

path <- "/software/cellranger-10.0.0/cellranger"

run_cellranger(command = command,
               input = fastq_paths,
               run.id = ids,
               reference = genome,
               sample = ids,
               cellranger = path,
               version = FALSE)
}
}
