% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_centrifuge.R
\name{run_centrifuge}
\alias{run_centrifuge}
\title{Run Centrifuge}
\usage{
run_centrifuge(
  input1 = NULL,
  input2 = NULL,
  index = NULL,
  sample.name = NULL,
  out.dir = NULL,
  threads = 10,
  min_hit_length = 15,
  parallel = FALSE,
  cores = 4,
  execute = TRUE,
  centrifuge = NULL,
  version = FALSE
)
}
\arguments{
\item{input1}{List of the paths to files containing to the forward reads}

\item{input2}{List of the paths to files containing to the reverse reads}

\item{index}{Path to the reference metagenome index}

\item{sample.name}{List of the sample names}

\item{out.dir}{Name of the directory from the Centrifuge output}

\item{threads}{Number of threads for Centrifuge to use, default set to 10}

\item{min_hit_length}{minimum length of partial hits, default set to the minimum allowed of 15}

\item{parallel}{Run in parallel, default set to FALSE}

\item{cores}{Number of cores/threads to use for parallel processing, default set to 4}

\item{execute}{Whether to execute the commands or not, default set to TRUE}

\item{centrifuge}{Path to the Centrifuge program}

\item{version}{Returns the version number}
}
\value{
A list with the Centrifuge commands
}
\description{
Runs the Centrifuge tool
}
\examples{
\dontrun{
path <- "/software/centrifuge-1.0.4/centrifuge"
in1 <- c("sample_1_1.fq","sample_2_1.fq")
in2 <- c("sample_1_2.fq","sample_2_2.fq")
sample_names <- c("sample_1","sample_2")
ref <- "/path/to/centrifuge/reference/index"

run_centrifuge(input1 = in1,
               input2 = in2,
               index = ref,
               sample.name = sample_names,
               out.dir = "test_dir",
               min_hit_length = 16,
               execute = FALSE,
               centrifuge = path)
}

}
