% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_fastq_screen.R
\name{run_fastq_screen}
\alias{run_fastq_screen}
\title{Run Fastq-Screen}
\usage{
run_fastq_screen(fq.files = NULL, out.dir = NULL, aligner = "bwa",
  conf = NULL, top = "500000,1000000", threads = 2,
  fastq_screen = NULL, version = FALSE)
}
\arguments{
\item{fq.files}{Vector of the full names and paths of the fastq files, usually only the forward reads set, required}

\item{out.dir}{Path to the directory to which to write the results. If NULL,
which is the default, a directory named "fastq_screen" is created in the current
working directory.}

\item{aligner}{The name of the program to perform the mapping, default bwa. Valid mappers are "bwa", "bowtie" and "bowtie2". Default set to bwa}

\item{conf}{The path to the configuration file, required}

\item{top}{Create a temporary datset by selecting the number of reads followed by how many to skip e.g deafaul is set to 500000,1000000 map 500 thousand reads and skip the first 1 million reads}

\item{threads}{The number of threads to use.  The default is 2.}

\item{fastq_screen}{The path to the fastq_screen executable.}

\item{version}{Returns the version number}
}
\value{
A file with the FastqScreen commands
}
\description{
Run the fastq-screen tool
}
\examples{
 \dontrun{
path <- "raw_reads"
mate1 <- list.files(path = path, pattern = "*_R1_001.fastq.gz$", full.names = TRUE)

fastq_screen_cmds <- run_fastq_screen(fq.files = mate1,
                                      out.dir = fastq.screen.dir,
                                      conf = "/export/jessie3/gmh5n/PipelineTest/fastq_screen.conf",
                                      top = "100000,5000",
                                      fastq_screen = "/software/fastq_screen_v0.13.0/fastq_screen")
}

}
