% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_fastqc.R
\name{run_fastqc}
\alias{run_fastqc}
\title{Run FastQC}
\usage{
run_fastqc(
  input = NULL,
  out.dir = out.dir,
  threads = NULL,
  parallel = FALSE,
  cores = 4,
  execute = TRUE,
  fastqc = NULL,
  version = FALSE
)
}
\arguments{
\item{input}{List of all of the fastq files, forward and reverse, required}

\item{out.dir}{Name of the directory to write the FastQC results, required}

\item{threads}{Number of threads for FastQC}

\item{parallel}{Run in parallel, default set to FALSE}

\item{cores}{Number of cores/threads to use for parallel processing, default set to 4}

\item{execute}{Whether to execute the commands or not, default set to TRUE}

\item{fastqc}{Path to the FastQC program, required}

\item{version}{Returns the version number}
}
\value{
A list with the FastQC commands
}
\description{
Runs the FastQC tool, a quality control tool for high throughput sequence data
}
\examples{
\dontrun{
fastqc.path <- "/software/FastQC-v0.11.8/fastqc"

# Version
fastqc.version <- ""
fastqc.version <- run_fastqc(fastqc = fastqc.path,
                             version = TRUE)
fastqc.version

# Run fastqc
reads.path <- "raw_reads"
mate1 <- list.files(path = reads.path, pattern = "*_R1_001.fastq.gz$", full.names = TRUE)
mate2 <- list.files(path = reads.path, pattern = "*_R2_001.fastq.gz$", full.names = TRUE)
# Merge the reads file names
all.reads <- c(mate1,mate2)

out.dir <- "fastqc"

fastqc.cmds <- run_fastqc(input = all.reads,
                          out.dir = out.dir,
                          fastqc = fastqc.path)
fastqc.cmds
}
}
