% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_flye.R
\name{run_flye}
\alias{run_flye}
\title{Run the Flye assembler}
\usage{
run_flye(
  input = NULL,
  sample_names = NULL,
  platform = NULL,
  out_dir = NULL,
  threads = 10,
  parallel = FALSE,
  cores = 4,
  execute = FALSE,
  flye = NULL,
  version = FALSE
)
}
\arguments{
\item{input}{List of the paths to files containing to the long reads, required}

\item{sample_names}{List of sample names, required}

\item{platform}{Names of the sequencing platform, choose either,
pacbio_raw, PacBio regular CLR reads less than 20 percent error
pacbio_corr, PacBio reads that were corrected with other methods less than 3 percent error
pacbio_hifi, PacBio HiFi reads
nano_raw, ONT regular reads, pre-Guppy5 less than 20 percent error
nano_corr, ONT reads that were corrected with other methods less than 3 percent error
nano_hq , ONT high-quality reads: Guppy5+ or Q20 less than 5 percent error}

\item{out_dir}{Name of the directory from the assembled reads}

\item{threads}{Number of threads for flye to use, default set to 10}

\item{parallel}{Run in parallel, default set to FALSE}

\item{cores}{Number of cores/threads to use for parallel processing, default set to 4}

\item{execute}{Whether to execute the commands or not, default set to TRUE}

\item{flye}{Path to the flye program, required}

\item{version}{Returns the version number}
}
\value{
A list with the Flye commands
}
\description{
Run the long read assembler Flye
}
\examples{

\dontrun{
path <- "/path/to/flye"
input_files <- c("sample1.fastq", "sample2.fastq")
sample_names <- gsub(".fastq","",input_files)
platform <- "nano_hq"
out_directory <- "Assemblies"

run_flye(flye = path,
         version = TRUE)

run_flye(input = input_files,
         sample_names = sample_names,
         platform = platform,
         out_dir = out_directory,
         threads = 10,
         parallel = FALSE,
         cores = 4,
         execute = FALSE,
         flye = path)
}

}
