% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_freebayes.R
\name{run_freebayes}
\alias{run_freebayes}
\title{Run Freebayes}
\usage{
run_freebayes(
  input = NULL,
  list = NULL,
  reference = NULL,
  output = NULL,
  ploidy = NULL,
  parallel = FALSE,
  cores = 4,
  execute = TRUE,
  freebayes = NULL,
  version = FALSE
)
}
\arguments{
\item{input}{List of aligned bam files}

\item{list}{File with a list of bam files}

\item{reference}{Reference genome sequence in fasta format}

\item{output}{List of file names for out put}

\item{ploidy}{Sets the default ploidy for the analysis, default st to 2}

\item{parallel}{Run in parallel, default set to FALSE}

\item{cores}{Number of cores/threads to use for parallel processing, default set to 4}

\item{execute}{Whether to execute the commands or not, default set to TRUE}

\item{freebayes}{Path to the freebayes program, required}

\item{version}{Returns the version number}
}
\value{
A list of freebayes commands
}
\description{
Run Freebayes
}
\examples{
\dontrun{
 freebayes <- "/usr/bin/freebayes"

 freebayes.verion <- run_freebayes.1(freebayes = freebayes,
                                   version = TRUE)

 freebayes.cmds <- run_freebayes(input = "input.bam",
                                 reference = "reference.fa",
                                 output = "out.vcf",
                                 ploidy = 2,
                                 execute = FALSE,
                                 freebayes = freebayes)
}

}
