% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hisat2.R
\name{run_hisat2}
\alias{run_hisat2}
\title{Run HISAT2}
\usage{
run_hisat2(
  input1 = NULL,
  input2 = NULL,
  index = NULL,
  sample.name = NULL,
  strandedness = NULL,
  no_splice = FALSE,
  known_splice = NULL,
  assembly = FALSE,
  phred = 33,
  threads = 10,
  out.dir = NULL,
  parallel = FALSE,
  cores = 4,
  execute = TRUE,
  hisat2 = NULL,
  version = FALSE
)
}
\arguments{
\item{input1}{List of the paths to files containing to the forward reads, required}

\item{input2}{List of the paths to files containing to the reverse reads, required for paired end sequence data}

\item{index}{Path to the reference genome hisat2 index, required}

\item{sample.name}{List of the sample names, required}

\item{strandedness}{Strand-specific information}

\item{no_splice}{Disable spliced alignment, use for Trypanosomes}

\item{known_splice}{File with known splice sites}

\item{assembly}{Reports alignments tailored for transcript assemblers}

\item{phred}{Quality score offsets, default set to illumina/sanger standard of 33}

\item{threads}{Number of threads for hisat2 to use, default set to 10}

\item{out.dir}{Name of the directory from the HISAT2 output. If NULL,
which is the default, a directory named "hisat2_alignments" is created in the current
working directory.}

\item{parallel}{Run in parallel, default set to FALSE}

\item{cores}{Number of cores/threads to use for parallel processing, default set to 4}

\item{execute}{Whether to execute the commands or not, default set to TRUE}

\item{hisat2}{Path to the HISAT2 program, required}

\item{version}{Returns the version number}
}
\value{
A list with the HISAT2 commands
}
\description{
Runs the HISAT2 tool, can be used for single end and paired end reads
}
\examples{
\dontrun{
 trimmed_reads_dir <- "trimmed_reads"
 mate1 <- list.files(path = trimmed_reads_dir, pattern = "*_R1_001.fastq$", full.names = TRUE)
 mate2 <- list.files(path = trimmed_reads_dir, pattern = "*_R2_001.fastq$", full.names = TRUE)

 sample_names <- unlist(lapply(strsplit(list.files(path = trimmed_reads_dir,
                        pattern = "*_R1_001.fastq$",
                        full.names = FALSE),"_"), `[[`, 1))

 index <- "/export/buzz1/Genome/Homo_sapiens/Ensembl/GRCH38_p7/Sequence/Transcriptome/
           KallistoIndex/GRCh38_p7.kall"

 # Paired End example
 strandedness <- "RF"
 hisat2.cmds <- run_hisat2(input1 = mate1,
                           input2 = mate2,
                           index = genome,
                           sample.name = sample.names,
                           strandedness = strandedness,
                           assembly = TRUE,
                           out.dir = hisat.alignments.dir,
                           hisat2 = "/software/hisat_v2-2.1.0/hisat2")

 # Single End example
 strandedness <- "R"
 hisat2.cmds <- run_hisat2(input1 = mate1,
                           index = genome,
                           sample.name = sample.names,
                           strandedness = strandedness,
                           assembly = TRUE,
                           out.dir = hisat.alignments.dir,
                           hisat2 = "/software/hisat_v2-2.1.0/hisat2")
 }

}
