% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_macs.R
\name{run_macs}
\alias{run_macs}
\title{Run Macs3}
\usage{
run_macs(
  command = NULL,
  treatment = NULL,
  control = NULL,
  format = NULL,
  genome.size = NULL,
  qvalue = NULL,
  broad.peaks = NULL,
  broad.peaks.cutoff = NULL,
  out.dir = NULL,
  experiment.name = NULL,
  parallel = FALSE,
  cores = 4,
  execute = TRUE,
  macs = NULL,
  version = FALSE
)
}
\arguments{
\item{command}{Tool arguments, choose form callpeak,bdgpeakcall,bdgbroadcall,
bdgcmp,bdgopt,cmbreps,bdgdiff,filterdup,predictd,pileup,randsample,refinepeak,callvar,hmmratac}

\item{treatment}{ChIP-seq treatment files list.}

\item{control}{Control files list.}

\item{format}{Format of the input files, choose from AUTO,BAM,SAM,BED,ELAND,ELANDMULTI,ELANDEXPORT,BOWTIE,BAMPE,BEDPE}

\item{genome.size}{Effective genome size.}

\item{qvalue}{Minimum FDR (q-value) cutoff for peak detection.}

\item{broad.peaks}{Boolean, if set to TRUE, MACS will try to call broad peaks}

\item{broad.peaks.cutoff}{Minimum FDR (q-value) cutoff for broad region.}

\item{out.dir}{Name of output directory}

\item{experiment.name}{Experiment name, which will be used to generate output file names.}

\item{parallel}{Run in parallel, default set to FALSE}

\item{cores}{Number of cores/threads to use for parallel processing, default set to 4}

\item{execute}{Whether to execute the commands or not, default set to TRUE}

\item{macs}{Path to the Macs3 program, required}

\item{version}{Returns the version number}
}
\description{
Macs3 is a spatial clustering approach for the identification
  of ChIP-enriched regions which was developed for calling narrow and broad peaks,
  for transcription factor binding or histone modifications from ChIP-seq data.
}
\examples{
\dontrun{
macs_path <- "/home/gmh5n/.localpython/bin/macs3"

macs_version <- run_macs(macs = macs_path,
                         version = TRUE)

macs_version

macs_command <- "callpeak"
treatment <- c("1_sorted.bam","2_sorted.bam")
experiment.name <- gsub("_sorted.bam","",treatment)
control <- c("1-input_sorted.bam","2-input_sorted.bam")
size <- 1.9e9
macs_dir <- "macs"
macs_commands <- run_macs(command = command,
                          treatment = treatment,
                          control = control,
                          format = "BAM",
                          genome.size = size,
                          broad.peaks = TRUE,
                          broad.peaks.cutoff = 0.05,
                          out.dir = macs_dir,
                          parallel = TRUE,
                          cores = 2,
                          execute = FALSE,
                          experiment.name = experiment.name,
                          macs = macs_path)
macs_commands
}
}
