% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mageck.R
\name{run_mageck}
\alias{run_mageck}
\title{Run Mageck}
\usage{
run_mageck(
  command = NULL,
  umi.list = NULL,
  fastq = NULL,
  sample.label = NULL,
  prefix = NULL,
  parallel = FALSE,
  cores = 4,
  execute = TRUE,
  version = NULL,
  mageck = NULL
)
}
\arguments{
\item{command}{Mageck command to run, currently only count is implemented}

\item{umi.list}{file contating the list of UMIs}

\item{fastq}{Trimmed sample files in fastq format, should only contain the umi sequence after trimming with cutadapt}

\item{sample.label}{Sample name}

\item{prefix}{Name to add to results files}

\item{parallel}{Run in parallel, default set to FALSE}

\item{cores}{Number of cores/threads to use for parallel processing, default set to 4}

\item{execute}{Whether to execute the commands or not, default set to TRUE}

\item{version}{Returns the version number}

\item{mageck}{Pathe to the mageck program, required}
}
\value{
List of Mageck commands
}
\description{
Run Mageck
}
\examples{
\dontrun{
mageck.version <- run_mageck(version = TRUE,
                            mageck = path)
mageck.version

command <- "count"
list <- "umi-list.tsv"
reads <- "MPRA_R1.fastq.gz"
label <- "MPRA"
prefix <- "MPRA"

mageck.cmds <- run_mageck(command = command,
                          umi.list = list,
                          fastq = reads,
                          sample.label = label,
                          prefix = prefix,
                          execute = FALSE,
                          mageck = path)

mageck.cmds

}
}
