% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_masurca.R
\name{run_masurca}
\alias{run_masurca}
\title{Run the MaSuRCA assembly program}
\usage{
run_masurca(
  input1 = NULL,
  input2 = NULL,
  input_long = NULL,
  output = NULL,
  threads = 10,
  parallel = FALSE,
  cores = 4,
  execute = TRUE,
  masurca = NULL,
  version = FALSE
)
}
\arguments{
\item{input1}{List of the paths to files containing to the forward reads, can be gzipped}

\item{input2}{List of the paths to files containing to the reverse reads, can be gzipped}

\item{input_long}{List of the paths to files containing the nanopore ot pacbio read, can be gzipped}

\item{output}{Name of assembly script, default assemble.sh}

\item{threads}{Number of threads}

\item{parallel}{Run in parallel, default set to FALSE}

\item{cores}{Number of cores/threads to use for parallel processing, default set to 4}

\item{execute}{Whether to execute the commands or not, default set to TRUE}

\item{masurca}{Path to the masurca program, required}

\item{version}{Returns the version number}
}
\value{
A list with the masurca commands
}
\description{
@description Creates a shell script to perform the hybrid assembly an correction of illumina and either nanopore or pacbio reads
}
\examples{
\dontrun{
# Path to masurca executable file
 path <- "/software/MaSuRCA-4.1.1/bin/masurca"

 # Check the program version
 run_masurca(masurca = path,
             version = TRUE)

 # Set paths to the reads files
 fastq1 <- "illumina1.fq.gz"
 fastq2 <- "illumina2.fq.gz"
 nanopore <- "nanopore.fq.gz"

 # Run the masurca commands to create the assembly script
 run_masurca(input1 = fastq1,
             input2 = fastq2,
             input_long = nanopore,
             threads = 40,
             masurca = path)
}

}
