% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_multiqc.R
\name{run_multiqc}
\alias{run_multiqc}
\title{Run MultiQC}
\usage{
run_multiqc(workingDir = NULL, multiqc = NULL, version = FALSE)
}
\arguments{
\item{workingDir}{Current working directory, parent directory for the logfiles, required}

\item{multiqc}{Path to the MultiQC program, required}

\item{version}{Returns the version number}
}
\description{
Run MultiQC
}
\examples{
 \dontrun{
 workingDir <- getwd()
 run_multiqc(workingDir,
             multiqc = "/usr/local/bin/multiqc")
 }
}
