% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nanofilt.R
\name{run_nanofilt}
\alias{run_nanofilt}
\title{Run the NanoFilt program}
\usage{
run_nanofilt(
  input = NULL,
  sample_names = NULL,
  out_dir = NULL,
  min_length = NULL,
  parallel = FALSE,
  cores = 4,
  execute = FALSE,
  nanofilt = NULL,
  version = FALSE
)
}
\arguments{
\item{input}{List of the paths to files containing to the nanopore reads, required}

\item{sample_names}{List of sample names, required}

\item{out_dir}{Name of the directory from the BWA output}

\item{min_length}{Minimum length for the reads}

\item{parallel}{Run in parallel, default set to FALSE}

\item{cores}{Number of cores/threads to use for parallel processing, default set to 4}

\item{execute}{Whether to execute the commands or not, default set to TRUE}

\item{nanofilt}{Path to the NanoFilt program, required}

\item{version}{Returns the version number}
}
\value{
A list with the NanoFilt commands
}
\description{
Runs the NanoFilt tool, can be used to filter nanopore reads reads
}
\examples{
 \dontrun{
path <- "/Path/To//NanoFilt"
input_files <- c("sample1.fastq", "sample2.fastq")
sample_names <- gsub(".fastq","",input_files)
out_directory <- "filtered_reads"
min_length <- 10000

run_nanofilt(input = input_files,
             sample_names = sample_names,
             out_dir = out_directory,
             min_length = min_length,
             nanofilt = path)
}


}
