% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_seqtk.R
\name{run_seqtk}
\alias{run_seqtk}
\title{Run Seqtk}
\usage{
run_seqtk(
  command = NULL,
  input = NULL,
  output = NULL,
  seed = NULL,
  reverse = NULL,
  num_reads = NULL,
  parallel = FALSE,
  cores = 4,
  execute = TRUE,
  seqtk = NULL
)
}
\arguments{
\item{command}{Seqtk command to run, at present can choose from 'seq' and 'sample', required}

\item{input}{List of fastq files, can be gzipped}

\item{output}{list of output file names, fastq format}

\item{seed}{Random seed for read selection}

\item{reverse}{Reverse complement the fastq sequences}

\item{num_reads}{Number of reads to output for sub sampling}

\item{parallel}{Run in parallel, default set to FALSE}

\item{cores}{Number of cores/threads to use for parallel processing, default set to 4}

\item{execute}{Whether to execute the commands or not, default set to TRUE}

\item{seqtk}{Path to the seqtk program, required}
}
\value{
A file with the Seqtk commands
}
\description{
Runs the seqtk tool
}
\examples{
\dontrun{
num_reads <- 1000
cmd <- "sample"
seed <- 23
reads <- c("reads1.fq.gz","reads2.fq.gz")
out <- c("reads1_out.fq.","reads2_out.fq")
seqtk_path <- "/usr/bin/seqtk"

run_seqtk(command = cmd,
          input = reads,
          output = out,
          seed = seed,
          num_reads = num_reads,
          parallel = TRUE,
          cores = 2,
          execute = FALSE,
          seqtk = seqtk_path)
}

}
