% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_snpeff.R
\name{run_snpeff}
\alias{run_snpeff}
\title{Run SnpEff}
\usage{
run_snpeff(
  command = NULL,
  input = NULL,
  output = NULL,
  genome = NULL,
  snpeff = NULL,
  execute = TRUE,
  version = FALSE
)
}
\arguments{
\item{command}{SnpEff command to run}

\item{input}{Input file, vcf format}

\item{output}{Name of output file, vcf format}

\item{genome}{Name of the genome}

\item{snpeff}{Path to the SnpEff program, required}

\item{execute}{Whether to execute the commands or not, default set to TRUE}

\item{version}{Returns the version number

 @examples
 \dontrun{
 snpeff.path <- "/software/snpEff-v5.0e/snpEff.jar"

 # Version
 res <- run_snpeff(snpeff = snpeff.path,
                   version = TRUE)

 # Download a premade genome
 command <- "download"
 genome <- "GRCh38.99"

 res <- run_snpeff(command = command,
                   genome = genome,
                   execute = FALSE,
                   snpeff = snpeff.path)

 # Run snpEff
 res <- run_snpeff(command = command,
                   genome = genome,
                   input = "input.vcf",
                   output = "annotated.vcf",
                   execute = FALSE,
                   snpeff = snpeff.path)
}}
}
\value{

}
\description{
Runs the SnpEff program.
}
