% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_snpsift.R
\name{run_snpsift}
\alias{run_snpsift}
\title{Run SnpSift}
\usage{
run_snpsift(
  command = NULL,
  input = NULL,
  output = NULL,
  filter = NULL,
  inverse = FALSE,
  dbsnp = NULL,
  parallel = FALSE,
  cores = 4,
  execute = TRUE,
  snpsift = NULL
)
}
\arguments{
\item{command}{SnpSift command to run, currently only fiter, annotate and vartype}

\item{input}{Input file, vcf format}

\item{output}{Name of output file, vcf format}

\item{filter}{Filter expression}

\item{inverse}{Show lines that do not match filter expression}

\item{dbsnp}{Path to the dbsnp file}

\item{parallel}{Run in parallel, default set to FALSE}

\item{cores}{Number of cores/threads to use for parallel processing, default set to 4}

\item{execute}{Whether to execute the commands or not, default set to TRUE}

\item{snpsift}{Path to the SnpSift program, required}
}
\value{
List of SnpSift commands
}
\description{
Runs the SnpSift program.
}
\examples{
\dontrun{
snpsift.path <- "/software/snpEff-v5.0e/SnpSift.jar"

run_snpsift(command = "filter",
            input = "file.vcf",
            output = "filtered.file.vcf",
            filter = "'( QUAL >= 30 )'",
            execute = FALSE,
            snpsift = snpsift.path)

run_snpsift(command = "annotate",
            input = "file.vcf",
            output = "annotated.file.vcf",
            dbsnp = "/datastore/GATK_resource_bundle/GRCh38/dbsnp_146.GRCh38.vcf.gz",
            execute = FALSE,
            snpsift = snpsift.path)

run_snpsift(command = "vartype",
            input = "file.vcf",
            output = "annotated.file.vcf",
            execute = FALSE,
            snpsift = snpsift.path)

}
}
