% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_spades.R
\name{run_spades}
\alias{run_spades}
\title{Run SPAdes tool}
\usage{
run_spades(
  input1 = NULL,
  input2 = NULL,
  out.dir = NULL,
  sample.name = NULL,
  isolate = FALSE,
  sc = FALSE,
  meta = FALSE,
  bio = FALSE,
  sewage = FALSE,
  corona = FALSE,
  rna = FALSE,
  plasmid = FALSE,
  metaviral = FALSE,
  metaplasmid = FALSE,
  rnaviral = FALSE,
  sanger = NULL,
  pacbio = NULL,
  nanopore = NULL,
  contigs = NULL,
  error_correction_only = NULL,
  assembly_only = NULL,
  parallel = FALSE,
  cores = 4,
  execute = TRUE,
  spades = NULL,
  version = FALSE
)
}
\arguments{
\item{input1}{List of the paths to files containing to the forward reads}

\item{input2}{List of the paths to files containing to the reverse reads}

\item{out.dir}{Name of the directory from the SPAdes output}

\item{sample.name}{List of the sample names}

\item{isolate}{This flag is highly recommended for high-coverage isolate and multi-cell data}

\item{sc}{This flag is required for MDA (single-cell) data}

\item{meta}{This flag is required for metagenomic data}

\item{bio}{This flag is required for biosyntheticSPAdes mode}

\item{sewage}{This flag is required for sewage mode}

\item{corona}{This flag is required for coronaSPAdes mode}

\item{rna}{This flag is required for RNA-Seq data}

\item{plasmid}{Runs plasmidSPAdes pipeline for plasmid detection}

\item{metaviral}{Runs metaviralSPAdes pipeline for virus detection}

\item{metaplasmid}{Runs metaplasmidSPAdes pipeline for plasmid detection in metagenomic datasets
(equivalent for meta & plasmid)}

\item{rnaviral}{This flag enables virus assembly module from RNA-Seq data}

\item{sanger}{Sanger sequence files}

\item{pacbio}{PacBIo sequence files}

\item{nanopore}{Nanopore sequence files}

\item{contigs}{Trusted cotig files}

\item{error_correction_only}{Run read error correction only}

\item{assembly_only}{Run assembly only}

\item{parallel}{Run in parallel, default set to FALSE}

\item{cores}{Number of cores/threads to use for parallel processing, default set to 4}

\item{execute}{Whether to execute the commands or not, default set to TRUE}

\item{spades}{Path to the SPAdes program}

\item{version}{Returns the version number}

\item{iontorrent}{This flag is required for IonTorrent data}
}
\value{
A list of SPAdes commands
}
\description{
Runs the SPAdes assembler
}
\examples{
 \dontrun{
 path <- "/software/SPAdes-4.0.0-Linux/bin/spades.py"

 run_spades(spades = path,
            version = TRUE)
 fastq1 <- c("sample1_R1.fq", "sample2_R1.fq")
 fastq2 <- c("sample1_R2.fq", "sample2_R2.fq")
 sample_names <- c("sample1", "sample2")
 run_spades(input1 = fastq1,
            input2 = fastq2,
            out.dir = "Out",
            isolate = TRUE,
            sample.name = sample_names,
            execute = FALSE,
            spades = path)
}
}
