% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_star.R
\name{run_star}
\alias{run_star}
\title{Run the Star program}
\usage{
run_star(
  mate1 = NULL,
  mate2 = NULL,
  genome = NULL,
  sample.name = NULL,
  out.dir = NULL,
  out.format = "BAM",
  unmapped = "Within",
  sam.attributes = "Standard",
  quant.mode = "GeneCounts",
  compressed = "zcat",
  threads = 10,
  parallel = FALSE,
  cores = 4,
  star = NULL,
  version = FALSE
)
}
\arguments{
\item{mate1}{List of the paths to files containing to the forward reads}

\item{mate2}{List of the paths to files containing to the reverse reads}

\item{genome}{Path to the reference genome index directory}

\item{sample.name}{List of the sample names}

\item{out.dir}{Name of the directory from the Star output}

\item{out.format}{Format of output file, default set to BAM}

\item{unmapped}{How to deal with unmapped reads, default set to "Within" which write unmapped reads to SAM/BAM file}

\item{sam.attributes}{Alignment attributes for the SAM/BAM file, default set to "Standard"}

\item{quant.mode}{Type of quantification required, default set to "GeneCounts"}

\item{compressed}{Compression mode for input reads files, default set to "zcat" for gzipped files, can use "bzcat" for bz2 files}

\item{threads}{Number of threads}

\item{parallel}{Run in parallel, default set to FALSE}

\item{cores}{Number of cores/threads to use for parallel processing, default set to 4}

\item{star}{Path to the Star program}

\item{version}{Returns the version number}
}
\value{
A list with the Star commands
}
\description{
Runs the Star alignment program, can be run in parallel on multiple cores.
}
\examples{
\dontrun{
path <- "/full/path/to/program"
genome <- "/full/path/to/genome"

mate1.trim <- List of paths to trimmed forward reads for alignment
mate2.trim <- List of paths to trimmed reverse reads for alignment
sample.names <- List os sample names

cmds <- run_star(mate1 = mate1.trim,
                 mate2 = mate2.trim,
                 genome = genome,
                 sample.name = sample.names,
                 out.dir = results.dir,
                 unmapped = "Within",
                 sam.attributes = "Standard",
                 quant.mode = "GeneCounts",
                 parallel = TRUE,
                 cores = 4,
                 star = path)

# Version number
version <- run_star(star = path,
                    version = TRUE)
}

}
