% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_stringtie.R
\name{run_stringtie}
\alias{run_stringtie}
\title{Run Stringtie}
\usage{
run_stringtie(
  input = NULL,
  threads = 10,
  trimming = TRUE,
  strandedness = NULL,
  reference.gtf = NULL,
  trans.estimate = FALSE,
  gene.estimate = FALSE,
  merge = FALSE,
  out = NULL,
  sample.name = NULL,
  ballgown = FALSE,
  parallel = FALSE,
  cores = 4,
  execute = TRUE,
  stringtie = NULL,
  version = FALSE
)
}
\arguments{
\item{input}{List of aligned files in sorted bam format, required}

\item{threads}{Number of threads for stringtie to use, default set to 10}

\item{trimming}{Disable trimming of predicted transcripts based on coverage, by default coverage trimming is enabled}

\item{strandedness}{Strand spcific reads, values are "first" or "second"}

\item{reference.gtf}{Path to the gtf file for guiding the assembly process}

\item{trans.estimate}{Estimates the abundance of given reference transcripts, requires the reference GTF}

\item{gene.estimate}{Estimates the abundance of given reference genes, requires the reference GTF}

\item{merge}{Assemble transcripts from multiple input files generating a unified non-redundant set of isoforms}

\item{out}{Name of the directory from the Kallisto output. If NULL,
which is the default, a directory named "stringtie" is created in the current working directory.}

\item{sample.name}{List of the sample names, required}

\item{ballgown}{Enable output of Ballgown table files which will be created in the
same directory as the output GTF (requires -G, -o recommended)}

\item{parallel}{Run in parallel, default set to FALSE}

\item{cores}{Number of cores/threads to use for parallel processing, default set to 4}

\item{execute}{Whether to execute the commands or not, default set to TRUE}

\item{stringtie}{Path to the stringtie program, required}

\item{version}{Returns the version number}
}
\value{
A list with the stringtie commands
}
\description{
Runs the stringtie tool
}
\examples{
 \dontrun{
# Version
stringtie.commands <- run_stringtie(stringtie = "/software/stringtie-1.3.6/stringtie",
                                    version = TRUE)
stringtie.commands

# Test command
sample_names <- unlist(lapply(strsplit(list.files(path = "trimmed_reads",
                       pattern = "*_R1_001.fastq$", full.names = FALSE),"_"), `[[`, 1))
gtf <- "/export/buzz1/Genome/Homo_sapiens/Ensembl/GRCh38_release_79/Sequence/Transcriptome/
         WholeTranscriptome/Homo_sapiens_GRCh38_79_primary_assembly_whole_transcriptome.gff"
bam.files <- list.files(path = hisat.alignments.dir, pattern = "sorted.bam$",
                        full.names = TRUE, recursive = TRUE)
stringtie.commands <- run_stringtie(input = bam.files,
                                    trimming = FALSE,
                                    strandedness = "first",
                                    reference.gtf = gtf,
                                    trans.estimate = TRUE,
                                    gene.estimate = TRUE,
                                    out = "stringtie",
                                    sample.name = sample_names,
                                    stringtie = "/software/stringtie-1.3.6/stringtie")
stringtie.commands
}

}
