% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_vcf_index.R
\name{run_vcf_index}
\alias{run_vcf_index}
\title{Compress and index vcf files}
\usage{
run_vcf_index(
  input = NULL,
  type = NULL,
  parallel = FALSE,
  cores = 4,
  execute = TRUE,
  bgzip = NULL,
  tabix = NULL
)
}
\arguments{
\item{input}{List of vcf files to index}

\item{type}{File type for indexing, e.g. "vcf"}

\item{parallel}{Run in parallel, default set to FALSE}

\item{cores}{Number of cores/threads to use for parallel processing, default set to 4}

\item{execute}{Whether to execute the commands or not, default set to TRUE}

\item{bgzip}{Path to the bgzip program, required}

\item{tabix}{Path to the tabix program, required}
}
\value{
A list with the indexing commands
}
\description{
Compress and index vcf files
}
\examples{
\dontrun{
bgzip.path <- "/software/htslib-1.9/bin/bgzip"
tabix.path <- "/software/htslib-1.9/bin/tabix"

vcf.files <- c("file1.vcf","file2.vcf")

file.type <- "vcf"

index.vcf.cmds <- run_vcf_index(input = vcf.files,
                                type = file.type,
                                execute = FALSE,
                                bgzip = bgzip.path,
                                tabix = tabix.path)
}

}
