% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_vcftools.R
\name{run_vcftools}
\alias{run_vcftools}
\title{Run vcftools}
\usage{
run_vcftools(
  input = NULL,
  output = NULL,
  plink = FALSE,
  execute = TRUE,
  vcftools = NULL,
  version = NULL
)
}
\arguments{
\item{input}{Path to the input vcf file}

\item{output}{Path to the output files}

\item{plink}{Create plink files, boolean, default st to FALSE}

\item{execute}{Whether to execute the commands or not, default set to TRUE}

\item{vcftools}{Path to the vcftools program, required}

\item{version}{Returns the version number}
}
\value{
Alist of vcftools commnads
}
\description{
Runs the vcftool, currently only set for creating PLINK Map and Ped files
}
\examples{
\dontrun{
vcftools.path <- "/software/vcftools-v0.1.16/bin/vcftools"

run_vcftools(vcftools = vcftools.path,
             version = TRUE)

input.file <- "variants.vcf"
output.file <- gsub("vcf","plink",input.file)

run_vcftools(input = input.file,
             output = output.file,
             plink = TRUE,
             execute = FALSE,
             vcftools = vcftools.path)
}

}
