% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_pheatmap.R
\name{save_pheatmap_png}
\alias{save_pheatmap_png}
\title{Save pheatmap}
\usage{
save_pheatmap_png(
  plot = NULL,
  filename = NULL,
  path = NULL,
  width = 1200,
  height = 1000,
  res = 150
)
}
\arguments{
\item{plot}{pheatmap oject}

\item{filename}{file name or file path name to write the pheatmap object}

\item{path}{Path of the directory to save plot}

\item{width}{Width of plot}

\item{height}{Height of plot}

\item{res}{Resolution of plot}
}
\value{
No return object
}
\description{
Save the pheatmap object as a png file
}
\examples{
\dontrun{

ph <- pheatmap(log2_mean_differential_genes

save_pheatmap_png(ph, "heatmap.png")
}

}
