% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_table.R
\name{save_table}
\alias{save_table}
\title{Save table to Typst source file}
\usage{
save_table(table, filename, create_dir = TRUE)
}
\arguments{
\item{table}{Character string}

\item{filename}{File name to create on disk (excluding path and extension)}

\item{create_dir}{Whether to create the output directory if it doesn't exist.}
}
\description{
Takes a text string specifying a Typst table, and writes it to a Typst file
in \verb{output/_tables/}.
}
\details{
Returns the path to the produced Typst source, for use in
\code{render_manuscript()}.
}
