% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odbc.R
\name{sendTable}
\alias{sendTable}
\title{sendTable}
\usage{
sendTable(
  email,
  driver = "ODBC Driver 17 for SQL Server",
  server,
  database,
  schema,
  tableName,
  dt,
  append = TRUE,
  overwrite = FALSE,
  newDataTypes = NULL
)
}
\arguments{
\item{email}{Whatever UID you sign into SQL server with.}

\item{driver}{Try leaving empty first.
You can use ```odbc::odbcListDrivers()``` to see which
drivers you have ifthe default doesn't work.}

\item{server}{The server}

\item{database}{Need to specify a connection to the database.}

\item{schema}{The schema you are sending the table to.}

\item{tableName}{The name the table will have in the database}

\item{dt}{The table in your R session you want to upload.}

\item{append}{Should it be unioned to a current table?}

\item{overwrite}{Should you overwrite an existing table. USE WITH CAUTION.}

\item{newDataTypes}{A named vector passed to field.types in DBI::dbWriteTable}
}
\value{
Nothing, this uploads a table to SQL server.
}
\description{
Upload a table to SQL Server
}
