% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subChars.R
\name{subChars}
\alias{subChars}
\title{subChars}
\usage{
subChars(pattern, sub, x)
}
\arguments{
\item{pattern}{The character(s) you want to replace.}

\item{sub}{The string you want to replace them with.}

\item{x}{The character vector you want to do the replacement on.}
}
\value{
Opens a csv with whatever default program is selected in a new window
}
\description{
removes a list of characters and replaces them with another.
}
