% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_render_typst.R
\name{tar_render_typst}
\alias{tar_render_typst}
\title{Figure image file output target}
\usage{
tar_render_typst(
  name,
  filename = "manuscript",
  values = list(),
  figs = list(),
  tbls = list(),
  appendix = character(),
  bibliography = "references.bib",
  template = "article",
  appendix_template = "appendix",
  tidy_eval = targets::tar_option_get("tidy_eval"),
  packages = targets::tar_option_get("packages"),
  library = targets::tar_option_get("library"),
  error = targets::tar_option_get("error"),
  memory = targets::tar_option_get("memory"),
  garbage_collection = isTRUE(targets::tar_option_get("garbage_collection")),
  deployment = targets::tar_option_get("deployment"),
  resources = targets::tar_option_get("resources"),
  storage = targets::tar_option_get("storage"),
  retrieval = targets::tar_option_get("retrieval"),
  cue = targets::tar_option_get("cue"),
  description = targets::tar_option_get("description")
)
}
\arguments{
\item{name}{Symbol, name stub for the target. Will be given prefixes \code{file_}
and \code{raw_} for the file and imported raw data targets, respectively.}

\item{filename}{String, filename stub for the saved image files.}

\item{tidy_eval}{Logical, whether to enable tidy evaluation when interpreting
command and pattern.}

\item{packages}{Character vector of packages to load right before the target
runs.}

\item{library}{Character vector of library paths to try when loading
packages.}

\item{error}{Character of length 1, what to do if the target
stops and throws an error. Options:
\itemize{
\item \code{"stop"}: the whole pipeline stops and throws an error.
\item \code{"continue"}: the whole pipeline keeps going.
\item \code{"null"}: The errored target continues and returns \code{NULL}.
The data hash is deliberately wrong so the target is not
up to date for the next run of the pipeline. In addition,
as of \code{targets} version 1.8.0.9011, a value of \code{NULL} is given
to upstream dependencies with \code{error = "null"} if loading fails.
\item \code{"abridge"}: any currently running targets keep running,
but no new targets launch after that.
\item \code{"trim"}: all currently running targets stay running. A queued
target is allowed to start if:
\enumerate{
\item It is not downstream of the error, and
\item It is not a sibling branch from the same \code{\link[targets:tar_target]{targets::tar_target()}} call
(if the error happened in a dynamic branch).
}

The idea is to avoid starting any new work that the immediate error
impacts. \code{error = "trim"} is just like \code{error = "abridge"},
but it allows potentially healthy regions of the dependency graph
to begin running.
(Visit \url{https://books.ropensci.org/targets/debugging.html}
to learn how to debug targets using saved workspaces.)
}}

\item{memory}{Character of length 1, memory strategy. Possible values:
\itemize{
\item \code{"auto"} (default): equivalent to \code{memory = "transient"} in almost
all cases. But to avoid superfluous reads from disk,
\code{memory = "auto"} is equivalent to \code{memory = "persistent"} for
for non-dynamically-branched targets that other targets
dynamically branch over. For example: if your pipeline has
\code{tar_target(name = y, command = x, pattern = map(x))},
then \code{tar_target(name = x, command = f(), memory = "auto")}
will use persistent memory for \code{x}
in order to avoid rereading all of \code{x}
for every branch of \code{y}.
\item \code{"transient"}: the target gets unloaded
after every new target completes.
Either way, the target gets automatically loaded into memory
whenever another target needs the value.
\item \code{"persistent"}: the target stays in memory
until the end of the pipeline (unless \code{storage} is \code{"worker"},
in which case \code{targets} unloads the value from memory
right after storing it in order to avoid sending
copious data over a network).
}}

\item{garbage_collection}{Logical: TRUE to run base::gc() just before the
target runs, in whatever R process it is about to run (which could be a
parallel worker). FALSE to omit garbage collection.}

\item{deployment}{Character of length 1. If deployment is "main", then the
target will run on the central controlling R process. Otherwise, if
deployment is "worker" and you set up the pipeline with
distributed/parallel computing, then the target runs on a parallel
worker.}

\item{resources}{Object returned by tar_resources() with optional settings
for high-performance computing functionality, alternative data storage
formats, and other optional capabilities of targets.}

\item{storage}{Character string to control when the output of the target
is saved to storage. Only relevant when using \code{targets}
with parallel workers (\url{https://books.ropensci.org/targets/crew.html}).
Must be one of the following values:
\itemize{
\item \code{"worker"} (default): the worker saves/uploads the value.
\item \code{"main"}: the target's return value is sent back to the
host machine and saved/uploaded locally.
\item \code{"none"}: \code{targets} makes no attempt to save the result
of the target to storage in the location where \code{targets}
expects it to be. Saving to storage is the responsibility
of the user. Use with caution.
}}

\item{retrieval}{Character string to control when the current target
loads its dependencies into memory before running.
(Here, a "dependency" is another target upstream that the current one
depends on.) Only relevant when using \code{targets}
with parallel workers (\url{https://books.ropensci.org/targets/crew.html}).
Must be one of the following values:
\itemize{
\item \code{"auto"} (default): equivalent to \code{retrieval = "worker"} in almost all
cases.
\item \code{"worker"}: the worker loads the target's dependencies.
\item \code{"main"}: the target's dependencies are loaded on the host machine
and sent to the worker before the target runs.
\item \code{"none"}: \code{targets} makes no attempt to load its
dependencies. With \code{retrieval = "none"}, loading dependencies
is the responsibility of the user. Use with caution.
}}

\item{cue}{An optional object from tar_cue() to customize the rules that
decide whether the target is up to date.}

\item{description}{Character of length 1, a custom free-form human-readable
text description of the target.}

\item{fn}{Function to read the data file into R.}

\item{...}{Additional arguments passed on to \code{fn}.}
}
\value{
A list of two \code{\link[targets:tar_target]{targets::tar_target()}} objects: one with 'file'
storage format (data file), one with standard target format (a
tibble/data frame).
}
\description{
Target factory for data import targets. Creates a pair of targets prefixed
\code{file_} and \code{raw_} for the data file and imported data objects,
respectively.
}
\details{
See \code{\link[targets:tar_target]{targets::tar_target()}} for details of all arguments other than
\code{filename} and those passed through \code{...}.
}
