% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toCSV.R
\name{toCSV}
\alias{toCSV}
\title{toCSV}
\usage{
toCSV(x, name = paste0(sample(LETTERS, 16, replace = TRUE), collapse = ""))
}
\arguments{
\item{x}{the data.table(frame) to write}

\item{name}{if you want to give it a custom name. Otherwise, 16 byte string is used.}
}
\value{
Opens a csv with whatever default program is selected in a new window
}
\description{
writes a data.table(frame) to a temp file
}
