% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_json.R
\name{write_to_json}
\alias{write_to_json}
\title{Write outputs to JSON file for Typst import}
\usage{
write_to_json(x, filename, ...)
}
\arguments{
\item{x}{List of R objects}

\item{path}{JSON filename stub}
}
\description{
Write outputs to JSON file for Typst import
}
