% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bowtie.R
\name{run_bowtie}
\alias{run_bowtie}
\title{Run bowtie}
\usage{
run_bowtie(
  mate1 = NULL,
  mate2 = NULL,
  index = NULL,
  sample.name = NULL,
  out.dir = NULL,
  unaligned = NULL,
  format = "SAM",
  parallel = FALSE,
  cores = 4,
  bowtie = NULL,
  version = FALSE
)
}
\arguments{
\item{mate1}{List of the paths to files containing to the forward reads, required}

\item{mate2}{List of the paths to files containing to the reverse reads, required for paired end sequence data}

\item{index}{Path to the reference transcriptome kallisto index, required}

\item{sample.name}{List of the sample names, required}

\item{out.dir}{Name of the directory from the Bowtie output. If NULL,
which is the default, a directory named "bowtie_alignments" is created in the current working directory.}

\item{unaligned}{Name of the directory for the unaligned reads, fastq formatted}

\item{format}{Format for the aligned reads, currently only SAM is supported}

\item{parallel}{Run in parallel, default set to FALSE}

\item{cores}{Number of cores/threads to use for parallel processing, default set to 4}

\item{bowtie}{Path to the bowtie proram, required}

\item{version}{Returns the version number}
}
\value{
A list with the bowtie commands
}
\description{
Runs the bowtie tool, currently set up primarily to get unaligned reads
}
\examples{
 \dontrun{
bowtie.path <- "/software/bowtie-v1.2.3/bin/bowtie"

bowtie.version <- run_bowtie(bowtie = bowtie.path,
                             version = TRUE)

trimmed_reads_dir <- "trimmed_reads"
mate1 <- list.files(path = trimmed_reads_dir, pattern = "*_R1_001.fastq$", full.names = TRUE)

index <- "path/to/index/indes.ebwt"

sample_names <- unlist(lapply(strsplit(list.files(path = trimmed_reads_dir,
                                                  pattern = "*_R1_001.fastq$",
                                                  full.names = FALSE),"_"), `[[`, 1))

unaligned <- "rRNA_removed_reads"
out.dir <- "/dev/null" # Send the reads that align to be deleted

bowtie.cmds <- run_bowtie(mate1 = mate1,
                          index = index, sample.name = sample_names,
                          unaligned = unaligned,
                          out.dir = out.dir,
                          bowtie = bowtie.path)
}

}
