% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cutadapt.R
\name{run_cutadapt}
\alias{run_cutadapt}
\title{Run Cutadapt}
\usage{
run_cutadapt(
  mate1 = NULL,
  mate2 = NULL,
  mate1.out = NULL,
  mate2.out = NULL,
  quality = 15,
  nextseq = FALSE,
  minimum = NULL,
  trim.only = FALSE,
  cut = NULL,
  adapter1 = NULL,
  adapter2 = NULL,
  cutadapt = NULL,
  version = FALSE
)
}
\arguments{
\item{mate1}{List of the paths to files containing to the forward reads}

\item{mate2}{List of the paths to files containing to the reverse reads}

\item{mate1.out}{List of paths to the files to write the trimmed forward reads}

\item{mate2.out}{List of paths to the files to write the trimmed reverse reads}

\item{quality}{The lower limit for the phred score}

\item{nextseq}{Was the sequence data generated on a NextSeq 500, trims dark cycle bases appearing as high-quality G bases}

\item{minimum}{The length at which a trimmed read will be discarded}

\item{trim.only}{Only keep reads that have had adapters ttrimmed}

\item{cut}{Remove the first 'n' bases form the 5' end of the forward read}

\item{adapter1}{Sequence for the adapter for the forward read}

\item{adapter2}{Sequence for the adapter for the reverse read}

\item{cutadapt}{Path to the Cutadapt program, required}

\item{version}{Returns the version number}
}
\value{
A file with the Cutadapt commands and creates a directory of adapter and quality trimmed reads
}
\description{
Run the Cutadapt tool to remove sequencing adapters and low quality bases.
}
\examples{
 \dontrun{
 # Version number
 run_cutadapt(cutadapt = cutadapt.path,
                         version = TRUE)

# Trimmed reads directory
trimmed.reads.dir <- "trimmed_reads"
#Create the directory for the trimmed reads
dir.create(trimmed.reads.dir, showWarnings = FALSE)

read1.pattern <- "*_R1_001.fastq.gz$"
read2.pattern <- "*_R2_001.fastq.gz$"

reads.path <- "/export/buzz2/gpfrawdata/nextseq01/FastQ/2019/NeilBulleid/
               1466_NS205_0325_MarieAnnPringle_20190313"

mate1 <- list.files(path = reads.path,
                    pattern = read1.pattern,
                    full.names = TRUE)
mate1.out <- paste(trimmed.reads.dir,(list.files(path = reads.path,
                                                 pattern = read1.pattern,
                                                 full.names = FALSE)), sep = "/")

mate2 <- list.files(path = reads.path,
                    pattern = read2.pattern,
                    full.names = TRUE)
mate2.out <- paste(trimmed.reads.dir,(list.files(path = reads.path,
                                                 pattern = read2.pattern,
                                                 full.names = FALSE)), sep = "/")

# Single end
run_cutadapt(mate1 = mate1,
             mate1.out = mate1.out,
             quality = 25,
             minimum = 17,
             trim.only = TRUE,
             cut = 1,
             adapter1 = "AGATCGGAAGAGCACACGTCT",
             cutadapt = cutadapt.path)

# Paired end
run_cutadapt(mate1 = mate1,
             mate2 = mate2,
             mate1.out = mate1.out,
             mate2.out = mate2.out,
             quality = 25,
             minimum = 17,
             trim.only = TRUE,
             cut = 1,
             adapter1 = "AGATCGGAAGAGCACACGTCT",
             adapter1 = "AGATCGGAAGAGCGTCGTGTAGGGAAAGAGTGTAGATCTCGGTGGTCGCCGTATCATT",
             cutadapt = cutadapt.path)
 }

}
