% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gatk.R
\name{run_gatk}
\alias{run_gatk}
\title{Run GATK}
\usage{
run_gatk(
  command = NULL,
  input = NULL,
  output = NULL,
  reference = NULL,
  normal.sample = NULL,
  intervals = NULL,
  known.sites = NULL,
  sample.map = NULL,
  erc = NULL,
  temp = NULL,
  batch = NULL,
  threads = NULL,
  database = NULL,
  bqsr = NULL,
  vqsr = NULL,
  tranches = NULL,
  resources = NULL,
  annotations = NULL,
  mode = NULL,
  tranches.file = NULL,
  sensitivity.filter = NULL,
  variant.index = TRUE,
  variant.type = NULL,
  select.method = NULL,
  filter.expression = NULL,
  filter.name = NULL,
  parallel = FALSE,
  cores = 4,
  execute = TRUE,
  gatk = NULL
)
}
\arguments{
\item{command}{GATK command to run, required}

\item{input}{List of sorted bam files, required}

\item{output}{List of output files or empty/non-existant directory}

\item{reference}{Path to the fasta formatted reference}

\item{normal.sample}{sample name of normal}

\item{intervals}{Path to the intrvals list file, usually the exome coordiantes file}

\item{known.sites}{List of paths to the files containing known polymorphic sites}

\item{sample.map}{Path to tab seperated file mapping sample names to the gvcf file}

\item{erc}{Mode for emitting reference confidence scores, can be "NONE", "BP_RESOLUTION" and "GVCF"}

\item{temp}{Path to temporary directory}

\item{batch}{Batch size fior number of readers open at once,GenomicsDBImport only}

\item{threads}{Number of threads for opening VCFs in batches}

\item{database}{Name of the database directory}

\item{bqsr}{List of base quality recalibration files}

\item{vqsr}{Varian recalibration file}

\item{tranches}{List of levels of truth sensitivity at which to slice the data}

\item{resources}{Pre defined list of sites for which to apply a prior probability of being correct}

\item{annotations}{List ofnames of the annotations to be used for calculations}

\item{mode}{Recalibration mode to employ, SNP or INDEL}

\item{tranches.file}{The input tranches file describing where to cut the data, from VariantRecalibrator}

\item{sensitivity.filter}{The truth sensitivity level at which to start filtering}

\item{variant.index}{Create a VCF index when writing a coordinate-sorted VCF file, boolean, default set to TRUE}

\item{variant.type}{Variant type to include in output, SNP or INDEL.}

\item{select.method}{Method to select filtered vaiants, to select only variant that pass all filters use 'vc.isNotFiltered()'}

\item{filter.expression}{String of filters and values}

\item{filter.name}{Name to identify the filtered variants}

\item{parallel}{Run in parallel, default set to FALSE}

\item{cores}{Number of cores/threads to use for parallel processing, default set to 4}

\item{execute}{Whether to execute the commands or not, default set to TRUE}

\item{gatk}{Path to the GATK suit of programs, required}
}
\value{
List of GATK commands
}
\description{
Runs the GATK suite of programs.
}
\examples{

\dontrun{
known.site.list <- c("Mills_and_1000G_gold_standard.indels.GRCh38.vcf.gz",
                      "1000G_phase1.snps.high_confidence.GRCh38.vcf.gz")

recalibration.files <- gsub(".bam","_recal_data.table",rg.bam.files)
command <-  "BaseRecalibrator"
fasta <- Path the genome fats
# BaseRecalibration
base.recalibration.cmds <- run_gatk(command = command,
                                    input = rg.bam.files,
                                    output = recalibration.files,
                                    reference = fasta,
                                    intervals = intervals.file,
                                    known.sites = known.site.list,
                                    execute = TRUE,
                                    gatk = gatk)
}

}
