% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gatk.R
\name{run_gatk}
\alias{run_gatk}
\title{Run GATK}
\usage{
run_gatk(
  command = NULL,
  input = NULL,
  output = NULL,
  reference = NULL,
  intervals = NULL,
  known.sites = NULL,
  bsqr = NULL,
  parallel = FALSE,
  cores = 4,
  execute = TRUE,
  gatk = NULL
)
}
\arguments{
\item{command}{GATK command to run, required}

\item{input}{List of sorted bam files, required}

\item{output}{List of output files}

\item{reference}{Path to the fasta formatted reference}

\item{intervals}{Path to the intrvals list file, usually the exome coordiantes file}

\item{known.sites}{List of paths to the files containing known polymorphic sites}

\item{bsqr}{List of base quality recalibration files}

\item{parallel}{Run in parallel, default set to FALSE}

\item{cores}{Number of cores/threads to use for parallel processing, default set to 4}

\item{execute}{Whether to execute the commands or not, default set to TRUE}

\item{gatk}{Path to the GATK suit of programs, required}
}
\value{
List of GATK commands
}
\description{
Runs the GATK suite of programs.
}
\examples{

\dontrun{
known.site.list <- c("Mills_and_1000G_gold_standard.indels.GRCh38.vcf.gz",
                      "1000G_phase1.snps.high_confidence.GRCh38.vcf.gz")

recalibration.files <- gsub(".bam","_recal_data.table",rg.bam.files)
command <-  "BaseRecalibrator"
fasta <- Path the genome fats
# BaseRecalibration
base.recalibration.cmds <- run_gatk(command = command,
                                    input = rg.bam.files,
                                    output = recalibration.files,
                                    reference = fasta,
                                    intervals = intervals.file,
                                    known.sites = known.site.list,
                                    execute = TRUE,
                                    gatk = gatk)
}

}
