% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_picard.R
\name{run_picard}
\alias{run_picard}
\title{Run Picard}
\usage{
run_picard(
  command = NULL,
  input = NULL,
  output = NULL,
  out.dir = NULL,
  refFlat = NULL,
  reference = NULL,
  rRNA.intervals = NULL,
  intervals = NULL,
  strand = NULL,
  remove.duplicates = FALSE,
  sample.name = NULL,
  library = NULL,
  platform = NULL,
  unit = NULL,
  vcf.files = NULL,
  parallel = FALSE,
  cores = 4,
  execute = TRUE,
  picard = NULL
)
}
\arguments{
\item{command}{Picard command to run, required}

\item{input}{List of sorted bam files, required}

\item{output}{List of output bam files, required for MarkDuplicates command}

\item{out.dir}{Name of the output directory, required}

\item{refFlat}{Path to the refFlat file, required for CollectRnaSeqMetrics commmand}

\item{reference}{Path to the fasta formatted reference for CollectWgsMetrics command}

\item{rRNA.intervals}{Path to the rRNAintrvals list file}

\item{intervals}{Path to the intrvals list file, usually the exome coordiantes file}

\item{strand}{Strand-specific information, "FR" for first strand or "RF" for reverse strand,
required for CollectRnaSeqMetrics commmand}

\item{remove.duplicates}{Set for removing marked duplicates, boolean default set to FALSE}

\item{sample.name}{List of the sample names, required}

\item{library}{Read group library}

\item{platform}{Sequencing platform, e.g. ILLUMINA}

\item{unit}{Platform unit, e.g. run barcode or number}

\item{vcf.files}{List of VCF files to sort and merge}

\item{parallel}{Run in parallel, default set to FALSE}

\item{cores}{Number of cores/threads to use for parallel processing, default set to 4}

\item{execute}{Whether to execute the commands or not, default set to TRUE}

\item{picard}{Path to the Picard program, required}
}
\value{
A list with the Picard commands
}
\description{
Runs the Picard program. Curently only works for CollectRnaSeqMetrics, CollectWgsMetrics & MarkDuplicates command
}
\examples{
\dontrun{
sorted.bam.files <- list.files(path = hisat2.alignments.dir, pattern = "sorted.bam$",
                               full.names = TRUE, recursive = TRUE)
sample_names <- unlist(lapply(strsplit(list.files(path = trimmed_reads_dir,
                       pattern = "*_R1_001.fastq$",
                       full.names = FALSE),"_"), `[[`, 1))
picard.dir <- "picard"
refFlat.file <- "refFlatBDGP6.txt"

run_picard(command = "CollectRnaSeqMetrics",
           input = sorted.bam.files,
           out.dir = picard.dir,
           refFlat = refFlat.file,
           strand = strandedness,
           sample.name = sample.names,
           picard = picard.path)
}

}
