% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_umitools.R
\name{run_umitools}
\alias{run_umitools}
\title{Run UMITools}
\usage{
run_umitools(
  command = NULL,
  parallel = FALSE,
  cores = 4,
  umitools = NULL,
  version = FALSE
)
}
\arguments{
\item{command}{Umitools command}

\item{parallel}{Run in parallel, default set to FALSE}

\item{cores}{Number of cores/threads to use for parallel processing, default set to 4}

\item{umitools}{Path to the Umitools program, required}

\item{version}{Returns the version number}
}
\value{
A file with the Umitools commands
}
\description{
UMITools are a set of tools for dealing with Unique Molecular Identifiers.
Runs the commands whitelist, extract, group, dedup, count and count_tab.
}
\examples{
 \dontrun{
 # Version number
 run_umitools(umitools = umitools.path,
                         version = TRUE)
}

}
